backend/requirements.txt
# =============================================================================
# ReconScope - Backend Dependencies
# =============================================================================
# Install: pip install -r requirements.txt
# =============================================================================

# --- Web Framework -----------------------------------------------------------
fastapi>=0.100
uvicorn[standard]>=0.25

# --- Task Queue --------------------------------------------------------------
celery>=5.3
redis>=5.0
flower>=2.0

# --- Database ----------------------------------------------------------------
sqlalchemy[asyncio]>=2.0
alembic>=1.12
asyncpg>=0.29

# --- HTTP & Networking -------------------------------------------------------
httpx>=0.25
dnspython>=2.4
python-whois>=0.9
websockets>=12.0
python-nmap>=0.7

# --- Report Generation -------------------------------------------------------
reportlab>=4.0

# --- Scheduling --------------------------------------------------------------
croniter>=1.3

# --- Data Validation & Settings ----------------------------------------------
pydantic>=2.0
pydantic-settings>=2.0
python-multipart>=0.0.6

# --- Testing -----------------------------------------------------------------
pytest>=7.0
pytest-asyncio>=0.21
pytest-cov>=4.0
fakeredis>=2.0
aiosqlite>=0.19
